pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "breadmind"
version = "0.3.0"
description = "AI Infrastructure Agent for K8s, Proxmox, OpenWrt"
license = "MIT"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.40.0",
    "openai>=1.50.0",
    "asyncpg>=0.30.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "pyyaml>=6.0",
    "jinja2>=3.1.0",
    "aiohttp>=3.11.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "duckduckgo-search>=7.0.0",
    "asyncssh>=2.17.0",
    "cryptography>=43.0.0",
    "celery>=5.4.0",
    "redis>=5.0.0",
    "fastembed>=0.4.0",
    "alembic>=1.13.0",
    "psycopg2-binary>=2.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0",
]
container = [
    "docker>=7.0",
]
messenger = [
    "twilio>=9.0",
    "google-api-python-client>=2.100.0",
    "google-auth-oauthlib>=1.2.0",
]
browser = [
    "playwright>=1.40.0",
]
embeddings = [
    "sentence-transformers>=3.0.0",
]
tokenizer = [
    "tiktoken>=0.7.0",
]
cli = [
    "rich>=13.0.0",
]

[project.scripts]
breadmind = "breadmind.main:main"
breadmind-uninstall = "breadmind.uninstall:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
